Added support for multi-word expressions in aspect keywords.
This was accomplished by adding an inner function to make_doc() that checks for keywords containing any common characters used to split tokens by the spacy tokenizer. If one of these characters are found in the keyword, the keyword gets added to the tokenizer as an exception that should not be split.