Enabled keyword matching for non-lower case keywords.
Aspect sentiment pipeline was failing to appropriately use multi-word
keywords as the spacy tokenizer is case sensitive, while the keyword
matcher uses only the lower case form of strings. This meant that:
    1. if a keyword like "United States" was being used, the tokenizier
       would only match verbatim instances (i.e. "United States" would
       be processed as a single token, while "united states" would be
       separated).
    2. my keyword matching code would convert all tokens to lower case,
       meaning that "United States" would be converted to
       "united states" and thus not match "United States" in the keyword
       list.

Fixed this problem by adjusting the logic of the keyword matching to
check for both literal strings and lower case versions.